MAINT: renamed utils
Renamed utils to be more descriptive.
