+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/stat_utils.c
@@ -29,7 +29,7 @@
 
 #include "mpfit.h"
 
-#include "utils.h"
+#include "stat_utils.h"
 #include "sort_rout.h"
 
 /**
